Refactor assignment management: restructure repository and service layers, update router endpoints, and improve dependency handling.
