+++ .gitignore
@@ -26,6 +26,10 @@ share/python-wheels/
 *.egg
 MANIFEST
 
+docker-compose.yml
+docker-entrypoint.sh
+Dockerfile-test
+
 # PyInstaller
 #  Usually these files are written by a python script from a template
 #  before PyInstaller builds the exe, so as to inject date/other infos into it.
