+++ app/core/deps.py
@@ -1,5 +1,5 @@
 from fastapi import Request
-from app.database.assignment import AssignmentRepo
+from app.database.assignment_repo import AssignmentRepo
 
 def get_repository(request: Request) -> AssignmentRepo:
     repo = getattr(request.app.state, "assignment_repo", None)
